Bug 1905211 part 1: Remove unnecessary script load in browser_text_spelling.js so that it can actually run. r=morgan
This test is disabled on CI because it's a bit fragile and we've never been able to work out how to fix that.
However, it's still useful locally.
Unfortunately, because it's disabled on CI, this problem wasn't caught when tests were moved around.

Differential Revision: https://phabricator.services.mozilla.com/D217062

UltraBlame original commit: 30bd5e3c1866066c13bf3ff2845d370081515de3