Bug 1867249, part 0: Replaced hash set with array for performance reasons. r=sefeng,dom-core
In the given code, all selections of all common inclusive ancestors are collected.
Typically, this will be only one.
In this case, looking up if a Selection pointer is already in the array boils down to one pointer comparison, which is much faster than inserting the pointer into the hashtable.
Since this is very hot code, changing the hashtable to an array already has a relatively significant influence.

Differential Revision: https://phabricator.services.mozilla.com/D205681

UltraBlame original commit: cb3c39200716ebaa1e12756227fcbf6a47f83997