Bug 1908249 - Get engine view in sync with navbar visibility r=android-reviewers,skhan
The navbar has an internal functionality of getting hidden whenever the keyboard is shown.
This would leave GeckoView still configured to show dynamic page elements in the space left
by considering both the address bar and the navbar heights and so would improperly place these.

The proposed fix will have the navbar inform when it's visibility depending on the keyboard
state changes so that it's integrators can sync this with the engine view.

Differential Revision: https://phabricator.services.mozilla.com/D218183

UltraBlame original commit: d735fac675d9ca646cfb93b6809d7feebea5d95b