Bug 1908279: Let the scope root search cross the shadow boundary if ::part selector is used. r=firefox-style-system-reviewers,emilio
This allows the scope search to extend beyond the shadow tree the element
is in any number of times, which is required to get to the DOM tree defining
the `::part` selector in question. If any shadow DOM tree in the chain does not
use `exportparts`, it is rejected when the style rule selector is matched.

Differential Revision: https://phabricator.services.mozilla.com/D218256

UltraBlame original commit: 13085ea1774da11e805086aa28cb7c77223a5d26