Bug 1911853 - Remove win 10 from try perf selector. r=perftest-reviewers,sparky
This patch prevents users from selecting tests available on Win 10. It
is temporarily enabled (Bug 1906951) for running sp3, but we only want
that to run in CI for data and not for users to run it.

Differential Revision: https://phabricator.services.mozilla.com/D218670

UltraBlame original commit: 8f93867d8a2e319d37d5c62284d1285100b48b5e