Bug 1630950 - [devtools] Parse colors for CSS variable declaration if computed value is a valid color. r=devtools-reviewers,jdescottes.
For the computation of options.supportsColor, we were checking
if the variable declaration value was a valid color. This is
working well for simple case, but not when the declaration is
using other variables.
Switching to the computed value fixes this case.
The output parser test for variables is modified so we can
check all the test cases for a css variable declaration (we only
use to test for a color declaration).
A couple test cases are skipped because they're not supported yet.
Bugs have been filed and reference added in comments.

Depends on D218519

Differential Revision: https://phabricator.services.mozilla.com/D218727

UltraBlame original commit: c30f0ff046b15e7669e14f2308201e9e6d09863d