Bug 1903454 - Initialize flags_ and fix operator& for RegExpFlags; r=iain
Upstream irregexp assumes that the flags_ will be initialized to zero
when checking for modifiers. Currently, this is not initialized, which is
why we saw inconsistent test failures between running locally and on try.
Adding an initializer means that RegExpFlags now has a non-trivial default
constructor, which is not permitted for a union, so I added a no-op
constructor to the formerly anonymous union that is a part of the Token
struct.

This also fixes operator& to return the result, rather than the lhs
argument.

Differential Revision: https://phabricator.services.mozilla.com/D219108

UltraBlame original commit: 89ed35e4a7a0462aaa748375efd8df6d52295b5c