Bug 1912229 - Move setPosition to sidebar init r=sidebar-reviewers,sclements
Previously we were only calling `setPosition` in `show()`. This meant
the "launcher" was not updated during session restore to be on the
correct side if "show on right" was set, because `SessionStore` only
calls `show()` if the sidebar panel is open.

To fix, `setPosition` has been moved to `init` instead.

Differential Revision: https://phabricator.services.mozilla.com/D219553

UltraBlame original commit: 8ea9693355974c2ec93da26125974b03098bc5c7