Bug 1907640 - part 5: set tier-up threshold based on function size.  r=rhunt.
Currently the tier-up threshold is set to 1.  This is overly aggressive and
largely negates any benefit from the lazy tiering.  This patch changes it to
`150 * (bodyLength ^ 1.5)`, a much more conservative setting which gives
demonstrable performance improvements.  This is an interim heuristic which is
likely to change as more benchmarking is carried out.

The power-of-1.5 aspect reflects the observation that Ion's compilation cost
tends to grow more than linearly with function size, especially if there are
many live ranges.

Differential Revision: https://phabricator.services.mozilla.com/D218405

UltraBlame original commit: 30d7a102e5707b7259dd062caf36c8b0928fd404