Bug 1879933 - Add a button to change the translation direction in about:translations. r=translations-reviewers,fluent-reviewers,desktop-theme-reviewers,bolsson,nordzilla,dao
Adds a button to the `about:translations` page that enables swapping the
selected source language and the selected target language. If there is populated
content when the languages are swapped, the current output text will become the
become the new source text, and a new output will be translated.

Differential Revision: https://phabricator.services.mozilla.com/D217045

UltraBlame original commit: 75e4425e79450a94273f084bae1b04f361d62649