Bug 1914673: Ensure WebRender provides update rects within the bounds of valid rects. r=gw
This is an extra-double-sure clipping of the scissor rect to the valid
rect. This shouldn't matter since the dirty rect is clipped before the
creation of either of the other two rects, but will ensure that rounding
issues from the application of the transform in `get_surface_rect` won't
leave us with  an out-of-bounds scissor rect.

Differential Revision: https://phabricator.services.mozilla.com/D219916

UltraBlame original commit: 1ca516749b3767be2054b1d1f7e2b3b0cb4830d9