Bug 1880582: Don't destroy the native window in nsCocoaWindow::Destroy. r=mstange
This is a temporary measure. We would like to destroy the native window
here, but we want to see the effect of removal on crash volume.

This patch also moves the call to BaseWindow::OnDestroy() to the end of
::Destroy(). This is necessary to get tests to pass. It may also be the
real fix for this Bug.

Differential Revision: https://phabricator.services.mozilla.com/D216066

UltraBlame original commit: 4b2edb2c26d4043e87dc4f88b22d1465afb85416