Bug 1913657 - assert fail in js::wasm::CompilePartialTier2 when Ion stack frame is too big.  r=bvisness.
wasm::CompilePartialTier2 calls various functions in ModuleGenerator to compile
a single function at tier 2.  It was previously release-asserting because it
assumed that those functions could only fail by OOMing.  But that
assumption/assertion is wrong.  In the test case in this bug, compilation fails
due to Ion reporting "stack frame too large", as a result of out-of-control
inlining.

Changes:

* wasm::CompilePartialTier2: new parameter `UniqueChars* error`, with which it
  can report to its caller, errors that occurred during compilation.  Remove
  release-assertion that OOM is the only possible error.

* Module::PartialTier2CompileTaskImp::runHelperThreadTask: this is the caller
  of the above function.  Add a new `UniqueChars error` to collect error(s),
  pass it to the above routine, and then hand those errors onwards to
  ReportTier2ResultsOffThread in case of compilation failure.

* unrelated drive-by change: completely remove class PartialTierModuleGenerator
  -- it is unused.

Currently it is impossible to use (a reduced version of) the supplied test case
as-is, because (1) we need to be able to cause a particular function to get
tiered up, which requires messing with tiering thresholds, and (2) subsequent
upadtes to our inlining heuristics should stop us ever generating such a huge
MIR function.

Differential Revision: https://phabricator.services.mozilla.com/D220386

UltraBlame original commit: 2238d0b663e868eab961752c3f8cffdbc31c9f80