Bug 1915291 - Keep track of stale Suggest suggestion types and ingest only stale types when their feature is updated. r=daisuke
This keeps track of the enabled suggestion types that have been ingested while
the app is running. I call those types "fresh". In the code, they're stored in
the `#ingestedSuggestionTypes` set. Types that have not been ingested and types
that are disabled are called "stale". When a feature is updated, we update the
staleness bookkeeping in `#ingestedSuggestionTypes` and ingest only enabled
stale types.

This is similar to how each feature has an `#isEnabled` property. It's our
cached value of the last-known state that we can compare to the current state.

Depends on D219943

Differential Revision: https://phabricator.services.mozilla.com/D220352

UltraBlame original commit: a619533788287b99a63ab49cb31276b856b4a09a