Bug 1915857 - [devtools] Compute waterfall ticks color in CSS. r=devtools-reviewers,ochameau.
We define the tick colors in CSS, using relative color syntax to
add the alpha we want.
We set the highlighter-* variables as registered properties so
the computed value will be the end color, even if we use
the light-dark function.

Differential Revision: https://phabricator.services.mozilla.com/D220706

UltraBlame original commit: 93ae1c65be47a03c6e3efefd18780bf753360374