Bug 1908699 - Removing temp file forced cleanup as windows 10 keeps having a file lock issue. r=gbeckley,jmoss
Removing the code and related tests that immediately deletes the remote-settings attachment bundle temp file.
This seems to cause an issue on windows 10 where it is locking the temp file for longer than we expect.
Given that this is a small temp file and will be cleaned up later anyway, this seems like an easy fix.

Differential Revision: https://phabricator.services.mozilla.com/D220737

UltraBlame original commit: 0a4378c2d7a8651c4d1e5856a10c7fc319556c53