Bug 1914202 - [mach-try-perf] Use Treeherder link for finding remote base/new revision. r=perftest-reviewers,fbilt
This patch changes the link to search for in mach-try-perf base/new revisions to the treeherder one instead of the hgmo link. This resolves an issue where the hg links would come in different styles depending on the number of commits made locally. This patch keeps the original hg matcher though as a fallback-method in case a treeherder link is unavailable. By default, we'll always use the treeherder revision even if an hg revision is found.

Differential Revision: https://phabricator.services.mozilla.com/D219770

UltraBlame original commit: 18a2ed0fd3f13f88cf0db6d66dd3435fa4cfaceb